+++ app/src/main/AndroidManifest.xml
@@ -34,7 +34,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".presentation.AutoClosingActivity"/>
+        <activity android:name=".presentation.ui.startup.AutoClosingActivity"/>
 
 
         <service android:name=".program.service.MyService">
@@ -45,7 +45,7 @@
         <receiver
             android:enabled="true"
             android:exported="true"
-            android:name=".presentation.StartupReceiver"
+            android:name=".presentation.ui.startup.StartupReceiver"
             android:permission="android.permission.RECEIVE_BOOT_COMPLETED">
 
             <intent-filter>
